BigDecimal field values must be rounded according to field format: minor refactoring #968
