+++ jmix-core/core/src/test/groovy/datatypes/DoubleDatatypeTest.groovy
@@ -18,17 +18,17 @@ package datatypes
 
 import format_strings.TestFormatStringsRegistry
 import io.jmix.core.metamodel.datatype.impl.DoubleDatatype
-import properties.TestCoreProperties
 import spock.lang.Specification
+import test_support.TestCoreProperties
 
 import java.text.ParseException
 
-class DoubleDatatypeTest  extends Specification {
+class DoubleDatatypeTest extends Specification {
 
     def "format/parse without locale, without rounding"() {
         def datatype = new DoubleDatatype()
         datatype.coreProperties = TestCoreProperties.builder()
-                .setDecimalValueRoundByFormat(false)
+                .setRoundDecimalValueByFormat(false)
                 .build()
 
         expect:
@@ -41,7 +41,7 @@ class DoubleDatatypeTest  extends Specification {
     def "format/parse without locale, with rounding"() {
         def datatype = new DoubleDatatype()
         datatype.coreProperties = TestCoreProperties.builder()
-                .setDecimalValueRoundByFormat(true)
+                .setRoundDecimalValueByFormat(true)
                 .build()
 
         expect:
@@ -55,7 +55,7 @@ class DoubleDatatypeTest  extends Specification {
         def datatype = new DoubleDatatype()
         datatype.formatStringsRegistry = new TestFormatStringsRegistry()
         datatype.coreProperties = TestCoreProperties.builder()
-                .setDecimalValueRoundByFormat(false)
+                .setRoundDecimalValueByFormat(false)
                 .build()
 
         expect:
@@ -71,7 +71,7 @@ class DoubleDatatypeTest  extends Specification {
         def datatype = new DoubleDatatype()
         datatype.formatStringsRegistry = new TestFormatStringsRegistry()
         datatype.coreProperties = TestCoreProperties.builder()
-                .setDecimalValueRoundByFormat(true)
+                .setRoundDecimalValueByFormat(true)
                 .build()
 
         expect:
